trie recovery works
